Rename git_here to git_retrieve
