+++ README.md
@@ -24,7 +24,7 @@ Some modules and plugins require external libraries. Please check the requiremen
 ### Modules
 Name | Description
 --- | ---
-[ansible.scm.git_here](https://github.com/ansible-collections/ansible.scm/blob/main/docs/ansible.scm.git_here_module.rst)|Retrieve a repository from a distant location and make it available on the execution node
+[ansible.scm.git_retrieve](https://github.com/ansible-collections/ansible.scm/blob/main/docs/ansible.scm.git_retrieve_module.rst)|Retrieve a repository from a distant location and make it available on the execution node
 [ansible.scm.git_publish](https://github.com/ansible-collections/ansible.scm/blob/main/docs/ansible.scm.git_publish_module.rst)|Publish changes from a repository available on the execution node to a distant location
 
 <!--end collection content-->
