+++ docs/ansible.scm.git_retrieve_module.rst
@@ -1,8 +1,8 @@
-.. _ansible.scm.git_here_module:
+.. _ansible.scm.git_retrieve_module:
 
 
 ********************
-ansible.scm.git_here
+ansible.scm.git_retrieve
 ********************
 
 **Retrieve a repository from a distant location and make it available on the execution node**
@@ -293,7 +293,7 @@ Examples
       gather_facts: true
       tasks:
         - name: Retrieve a repository from a distant location and make it available locally
-          ansible.scm.git_here:
+          ansible.scm.git_retrieve:
             origin:
               url: git@github.com:cidrblock/scm_testing.git
             upstream:
