+++ plugins/modules/git_retrieve.py
@@ -13,7 +13,7 @@ __metaclass__ = type
 # pylint: enable=invalid-name
 
 DOCUMENTATION = """
-module: git_here
+module: git_retrieve
 short_description: >-
   Retrieve a repository from a distant location and make it available on the execution node
 version_added: "1.0.0"
@@ -122,7 +122,7 @@ EXAMPLES = r"""
   gather_facts: true
   tasks:
     - name: Retrieve a repository from a distant location and make it available locally
-      ansible.scm.git_here:
+      ansible.scm.git_retrieve:
         origin:
           url: git@github.com:cidrblock/scm_testing.git
         upstream:
